+++ packages/platform-web/src/shuvi-app/react/__tests__/dynamic.test.tsx
@@ -2,15 +2,9 @@
  * @jest-environment jsdom
  */
 
-global.setImmediate =
-  global.setImmediate ||
-  ((fn: any, ...args: any) => global.setTimeout(fn, 0, ...args));
-global.clearImmediate =
-  global.clearImmediate || ((id: any) => global.clearTimeout(id));
-
 import * as React from 'react';
 import { getByText, render, waitFor, cleanup } from '@testing-library/react';
-import { wait } from 'shuvi-test-utils';
+import { wait } from 'shuvi-test-utils/shared';
 import dynamic from '../dynamic';
 
 describe('dynamic', () => {
