+++ packages/service/src/server/http-server/__tests__/router.test.ts
@@ -1,4 +1,4 @@
-import { findPort } from 'shuvi-test-utils';
+import { findPort } from 'shuvi-test-utils/findPort';
 import got from 'got';
 import { Server } from '../server';
 import { Router, getRouter } from '../router';
